Switch from Pods to Carthage, renamed Plex
This removes any traces of Cocoapods from the project and switches to
Carthage for dependency management.

Also updates SocketRocket framework to 0.4.2 in the process. Tried to
update to their 0.5.1, however it does not build properly. See issue:
https://github.com/facebook/SocketRocket/issues/474

Rename Plex to Plex Home Theater in preparation of adding Plex Media
Player client support.